requirements.txt
coldfront
djangorestframework
mozilla-django-oidc
django-scim2 >=0.19.0, <0.20.0
